+++ libsolc/libsolc.h
@@ -37,7 +37,7 @@ extern "C" {
 /// Callback used to retrieve additional source files.
 ///
 /// "Returns" two pointers that should be heap-allocated and are free'd by the caller.
-typedef void (*CStyleReadFileCallback)(char const* _path, char** o_contents, char** o_error);
+typedef void (*CStyleReadFileCallback)(void* _context, char const* _path, char** o_contents, char** o_error);
 
 /// Returns the complete license document.
 ///
@@ -53,7 +53,7 @@ char const* solidity_version() SOLC_NOEXCEPT;
 /// a "Standard Output JSON". Both are to be UTF-8 encoded.
 ///
 /// The pointer returned must not be freed by the caller.
-char const* solidity_compile(char const* _input, CStyleReadFileCallback _readCallback) SOLC_NOEXCEPT;
+char const* solidity_compile(char const* _input, CStyleReadFileCallback _readCallback, void* _readContext) SOLC_NOEXCEPT;
 
 /// Frees up any allocated memory.
 ///