+++ test/libsolidity/LibSolc.cpp
@@ -58,7 +58,7 @@ bool containsError(Json::Value const& _compilerResult, string const& _type, stri
 
 Json::Value compile(string const& _input, CStyleReadFileCallback _callback = nullptr)
 {
-	string output(solidity_compile(_input.c_str(), _callback));
+	string output(solidity_compile(_input.c_str(), _callback, nullptr));
 	Json::Value ret;
 	BOOST_REQUIRE(jsonParseStrict(output, ret));
 	solidity_free();
@@ -137,7 +137,7 @@ BOOST_AUTO_TEST_CASE(with_callback)
 	)";
 
 	CStyleReadFileCallback callback{
-		[](char const* _path, char** o_contents, char** o_error)
+		[](void* _context, char const* _path, char** o_contents, char** o_error)
 		{
 			// Caller frees the pointers.
 			if (string(_path) == "found.sol")